requirements.txt
pandas>=2.0.0
pymongo>=4.0.0
pymongo[srv]>=4.0.0
python-docx>=0.8.11
python-dateutil>=2.8.2 